audio: control: fix tv board set hdmitx mute control cause crash issue [1/1]
PD#SWPL-832

Problem:
kernel crash when hdmitx is not resigtered

Solution:
audio hal has a interface to control the hdmitx audio mute/unmute,
but for tv board, the hdmitx module is not registered in kernel.
the mixer api will cause kernel exception of null pointer.
So only apply that when in OTT module.

Verify:
not yet.

Change-Id: I1f27e771db654745c924a96909d05ee5f142ad06
Signed-off-by: Jian Xu <jian.xu@amlogic.com>